Extract runtime session and skill helpers
